// Hippocamp.Web/src/app/shared/generated/model/custom-page-dto.ts
/**
 * Hippocamp.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { MenuItemDto } from '././menu-item-dto';

export class CustomPageDto { 
    readonly IsEmptyContent?: boolean;
    CustomPageID?: number;
    CustomPageDisplayName?: string;
    CustomPageVanityUrl?: string;
    CustomPageContent?: string;
    MenuItem?: MenuItemDto;
    SortOrder?: number;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}